// src/app/enums/article-action-types.ts
export enum ArticleActionTypes {

  LOAD_ARTICLES = '[Article] Load Articles',
  LOAD_ARTICLES_SUCCESS = '[Article] Load Articles Success',
  LOAD_ARTICLES_FAILURE = '[Article] Load Articles Failure',

  ADD_ARTICLE = '[Article] Save',
  ADD_ARTICLE_SUCCESS = '[Article] Save Success',
  ADD_ARTICLE_FAILURE = '[Article] Save Failure',

  UPDATE_ARTICLE = '[Article] Update Article',
  UPDATE_ARTICLE_SUCCESS = '[Article] Update Article Success',
  UPDATE_ARTICLE_FAILURE = '[Article] Update Article Failure',

  DELETE_ARTICLE = '[Article] Delete Article',
  DELETE_ARTICLE_SUCCESS = '[Article] Delete Article Success',
  DELETE_ARTICLE_FAILURE = '[Article] Delete Article Failure',

  LOAD_ARTICLE = '[Article] Load Article',
  LOAD_ARTICLE_SUCCESS = '[Article] Load Article Success',
  LOAD_ARTICLE_FAILURE = '[Article] Load Article Failure',

}